onnx: set Protobuf_INCLUDE_DIR so native protobuf isn't picked up
In case native compilers and target compilers support a different
cpp standard, the abseil wrappers can be slightly different. Sometimes
this is obvious at link time, but for some inline functions it won't
be obvious until runtime.

This change sets Protobuf_INCLUDE_DIR to ensure that the correct
target header include dir is used when we build.